+++ tasks.py
@@ -30,6 +30,7 @@ ns.configure(
         "repository": "123456789012.dkr.ecr.us-east-1.amazonaws.com/myproject",
         "run": {
             "echo": True,
+            "pty": True,
         }
     }
 )